requirements.txt
Flask==2.0.0
gunicorn==21.2.0
MarkupSafe==2.1.0
Pillow==10.1.0
pytesseract== 0.3.10

